refactor: split the kanban board's lane geometry out of the page
kanban-page.tsx crossed the 500-line refactor prompt. The seam is "does a
card fit" versus "what is on the card": kanban-board.tsx now owns the lane
count, the lane frame and scroll region, the narrow single-lane strip and
the width floor, while the page keeps the boards, the selection, and every
dialog and mutation behind enter/n/d.

The only thing crossing the seam is the measured board width. 534 -> 398
plus a 199-line sibling; render track green and byte-identical output.
