+++ packages/kobe/src/tui-react/component/kanban-page.tsx
@@ -12,6 +12,12 @@
  * Enter opens {@link IssueDetailDialog}, whose Start hands an
  * {@link IssueChatStart} to the host). Column math is framework-free
  * (`state/issue-board.ts`): done > parked > linked-task > backlog.
+ *
+ * This file owns the DATA and the ACTIONS — which boards exist, what is
+ * selected, and every dialog/mutation behind `enter` / `n` / `d`. How many
+ * lanes fit and how one is drawn is {@link KanbanBoard}'s job
+ * (`component/kanban-board.tsx`); the only thing crossing that seam is the
+ * measured board width.
  */
 
 import { type BoxRenderable, TextAttributes } from "@opentui/core"
@@ -22,7 +28,7 @@ import type { RemoteOrchestrator, TaskEngineState } from "../../client/remote-or
 import { availableEngineIds } from "../../engine/account-detect"
 import { engineDisplayName } from "../../engine/interactive-command"
 import { errorMessage } from "../../lib/error-message"
-import { type BoardColumnKey, applyBoardAttention, buildIssueBoard, moveBoardSelection } from "../../state/issue-board"
+import { applyBoardAttention, buildIssueBoard, moveBoardSelection } from "../../state/issue-board"
 import { sidebarProjectLabel } from "../../tui/panes/sidebar/groups"
 import type { VendorId } from "../../types/task"
 import { useNotifications } from "../context/notifications"
@@ -33,38 +39,12 @@ import { isNarrowWidth } from "../lib/narrow-mode"
 import { useCursorFollow } from "../lib/use-cursor-follow"
 import { useDialog } from "../ui/dialog"
 import { DialogConfirm } from "../ui/dialog-confirm"
-import { FRAME } from "../ui/frame"
 import { quickForkDefaultVendor } from "../workspace/quick-fork"
 import type { IssueChatStart } from "../workspace/use-issue-chat"
 import { IssueDetailDialog } from "./issue-detail-dialog"
-import { KanbanCard } from "./kanban-card"
+import { KanbanBoard, needsSingleLane } from "./kanban-board"
 import { useKanbanBoards } from "./use-kanban-boards"
 
-const COLUMN_LABEL_KEY: Record<BoardColumnKey, string> = {
-  backlog: "kanban.column.backlog",
-  in_progress: "kanban.column.inProgress",
-  parked: "kanban.column.parked",
-  done: "kanban.column.done",
-}
-
-/**
- * Board-local width floor for the four-lane layout — deliberately NOT in
- * `lib/narrow-mode.ts`. That module's 70 columns is a WHOLE-TERMINAL
- * predicate (does the three-pane desktop layout fit at all); this one asks a
- * narrower question the sidebar's width also answers to: does a lane still
- * hold a readable card. At 100 terminal columns the desktop layout is fine
- * and the board still had 9 cells of card content.
- *
- * A card's own content needs 12: `issue.created` is `YYYY-MM-DD` at 10 cells
- * and the activity badge shares that row. LANE_CHROME is what stands between
- * the board's width and that content, per lane — lane border 2 + lane padding
- * 2 + the scrollbar gutter 1 + card border 2 + card padding 2.
- */
-const MIN_CARD_CELLS = 12
-const LANE_CHROME = 9
-const BOARD_LANES = 4
-/** 4 × (12 + 9) + 3 single-cell gaps = 87. */
-const MIN_BOARD_CELLS = BOARD_LANES * (MIN_CARD_CELLS + LANE_CHROME) + (BOARD_LANES - 1)
 /** paddingLeft + paddingRight on the page root, whose width is what we measure. */
 const PAGE_PADDING_CELLS = 4
 
@@ -85,8 +65,7 @@ export function KanbanPage(props: {
    *  opens already pointing at the work the task belongs to. */
   focusTask?: { readonly id: string; readonly repo: string }
 }): ReactNode {
-  const { theme, transparentBackground } = useTheme()
-  const columnBorder = transparentBackground ? theme.border : theme.borderSubtle
+  const { theme } = useTheme()
   const t = useT()
   const dialog = useDialog()
   // Surface mutation failures as on-screen toasts: under an alternate screen
@@ -106,8 +85,7 @@ export function KanbanPage(props: {
   // oscillate). `onSizeChange` + a ref is the tab-strip's pattern.
   const pageRef = useRef<BoxRenderable | null>(null)
   const [pageCells, setPageCells] = useState<number | null>(null)
-  const boardCells = pageCells === null ? null : pageCells - PAGE_PADDING_CELLS
-  const singleLane = boardCells === null ? narrow : boardCells < MIN_BOARD_CELLS
+  const singleLane = needsSingleLane(pageCells === null ? null : pageCells - PAGE_PADDING_CELLS) ?? narrow
 
   // Detected engines for the detail drawer's picker — one probe per page
   // open (account files on disk; cheap and refreshed enough).
@@ -336,33 +314,6 @@ export function KanbanPage(props: {
     ],
   }))
 
-  const columnAccent = {
-    backlog: theme.textMuted,
-    in_progress: theme.accent,
-    parked: theme.warning,
-    done: theme.success,
-  } satisfies Record<BoardColumnKey, unknown>
-
-  function card(issue: Issue, column: BoardColumnKey): ReactNode {
-    // Linked cards in the live lanes track their task's engine activity (the
-    // stay-on-the-board half of the background-start trigger); only In
-    // progress floats/counts the badge, Parked keeps it as passive signal.
-    const live = column === "in_progress" || column === "parked"
-    const activity = live && issue.taskId ? props.engineStates?.get(issue.taskId)?.state : undefined
-    return (
-      <KanbanCard
-        key={issue.id}
-        issue={issue}
-        column={column}
-        selected={issue.id === selectedId}
-        activity={activity}
-        onSelect={() => setSelectedId(issue.id)}
-        onOpen={() => openDetail(issue)}
-        boxRef={follow.rowRef(issue.id)}
-      />
-    )
-  }
-
   /** One-line rolling project selector — tab/←/→ (or click) cycles, no tab
    *  row. Label stays flush with the page's left edge. The full path is a
    *  wide-layout nicety; narrow keeps only the label that identifies. */
@@ -391,102 +342,6 @@ export function KanbanPage(props: {
             {t("kanban.empty")}
           </text>
         ) : null}
-      </box>
-    )
-  }
-
-  function column(col: (typeof columns)[number], opts?: { header?: boolean }): ReactNode {
-    return (
-      <box
-        key={col.key}
-        flexGrow={1}
-        flexBasis={0}
-        // Rounded like every other framed surface — see ui/frame.ts for why
-        // this is spread rather than written out.
-        {...FRAME}
-        borderColor={columnBorder}
-        paddingLeft={1}
-        paddingRight={1}
-      >
-        {(opts?.header ?? true) ? (
-          <box flexDirection="row" justifyContent="space-between">
-            <text fg={columnAccent[col.key]} attributes={TextAttributes.BOLD} wrapMode="none">
-              {t(COLUMN_LABEL_KEY[col.key])} ({col.issues.length + col.hiddenCount})
-            </text>
-            {col.key === "in_progress" && attentionCount > 0 ? (
-              <text fg={theme.warning} attributes={TextAttributes.BOLD} wrapMode="none">
-                {t("kanban.attention", { count: String(attentionCount) })}
-              </text>
-            ) : null}
-          </box>
-        ) : null}
-        {/* paddingRight keeps a one-cell gutter under the scrollbar thumb —
-            without it the thumb paints over the cards' right borders. The
-            horizontal bar is hidden outright: a lane never scrolls sideways. */}
-        <scrollbox
-          ref={follow.scrollRef}
-          flexGrow={1}
-          paddingTop={1}
-          paddingRight={1}
-          verticalScrollbarOptions={{ showArrows: false, trackOptions: { foregroundColor: "transparent" } }}
-          horizontalScrollbarOptions={{ visible: false }}
-        >
-          {col.issues.map((issue) => card(issue, col.key))}
-          {col.issues.length === 0 && col.hiddenCount === 0 ? (
-            <text fg={theme.textMuted} wrapMode="none">
-              {t("kanban.columnEmpty")}
-            </text>
-          ) : null}
-          {col.hiddenCount > 0 ? (
-            <text fg={theme.textMuted} wrapMode="none">
-              {t("kanban.more", { count: String(col.hiddenCount) })}
-            </text>
-          ) : null}
-        </scrollbox>
-      </box>
-    )
-  }
-
-  /** Narrow: one full-width lane (the selection's column) under a strip of
-   *  the other lanes' counts; ←/→ moves selection across lanes, and the
-   *  visible column follows it. Clicking a lane jumps to its first card. */
-  function narrowBoard(): ReactNode {
-    const active =
-      columns.find((col) => col.issues.some((issue) => issue.id === selectedId)) ??
-      columns.find((col) => col.issues.length > 0) ??
-      columns[0]
-    if (!active) return null
-    return (
-      <box flexDirection="column" flexGrow={1} paddingTop={1}>
-        <box flexDirection="row" gap={2}>
-          {columns.map((col) => (
-            <text
-              key={col.key}
-              fg={col.key === active.key ? columnAccent[col.key] : theme.textMuted}
-              attributes={col.key === active.key ? TextAttributes.BOLD : undefined}
-              wrapMode="none"
-              onMouseUp={() => {
-                const first = col.issues[0]
-                if (first) setSelectedId(first.id)
-              }}
-            >
-              {t(COLUMN_LABEL_KEY[col.key])} ({col.issues.length + col.hiddenCount})
-            </text>
-          ))}
-        </box>
-        {/* The strip above already names the active lane — the in-column
-            header would repeat it one row later. Blocked cards still read:
-            the attention float pins them to the top with warning borders. */}
-        {column(active, { header: false })}
-      </box>
-    )
-  }
-
-  function board(): ReactNode {
-    if (singleLane) return narrowBoard()
-    return (
-      <box flexDirection="row" gap={1} flexGrow={1} paddingTop={1}>
-        {columns.map((col) => column(col))}
       </box>
     )
   }
@@ -526,7 +381,16 @@ export function KanbanPage(props: {
       ) : (
         <>
           {projectSelector(activeBoard)}
-          {board()}
+          <KanbanBoard
+            columns={columns}
+            attentionCount={attentionCount}
+            selectedId={selectedId}
+            singleLane={singleLane}
+            {...(props.engineStates ? { engineStates: props.engineStates } : {})}
+            follow={follow}
+            onSelect={setSelectedId}
+            onOpen={openDetail}
+          />
         </>
       )}
     </box>
